cloner/requirements.txt
GitPython==3.0.5
pika